Update SSH keys
Add a passphrase to all SSH keys. To keep things ergonomic they are automatically unlocked on boot. This means if someone were to gain access to the system they could use the key on the machine but they could not copy it since there is no way to get the decrypted key out of the SSH agent (without root privileges). I consider the attack surface to be minimal and this to be a nice improvement.